# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PESkinPro Backend API",
    "version": "1.0.0",
    "description": "API Documentation for PESkinPro E-Commerce Backend",
    "contact": {
      "name": "Developer"
    }
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current Server (Relative)"
    },
    {
      "url": "http://localhost:5000/api/v1",
      "description": "API Server (Configured)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/vouchers": {
      "get": {
        "summary": "Daftar Semua Voucher Aktif",
        "description": "Mengambil semua kupon atau voucher yang tersedia dan masih bisa digunakan oleh pelanggan.",
        "tags": [
          "Order & Checkout"
        ],
        "responses": {
          "200": {
            "description": "Daftar voucher berhasil diambil.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "description": "Kode Voucher"
                      },
                      "title": {
                        "type": "string"
                      },
                      "subtitle": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string",
                        "enum": [
                          "shipping",
                          "promo"
                        ]
                      },
                      "savingLabel": {
                        "type": "string"
                      },
                      "validTo": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Buat Voucher Baru (Admin)",
        "description": "Administrator dapat membuat voucher promo atau ongkir baru.",
        "tags": [
          "Order & Checkout"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "title",
                  "type"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "PROMO50"
                  },
                  "title": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "shipping",
                      "promo"
                    ]
                  },
                  "discount_value": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Voucher berhasil dibuat."
          }
        }
      }
    },
    "/vouchers/{id}": {
      "get": {
        "summary": "Detail Voucher by ID",
        "description": "Mengambil informasi rinci sebuah voucher berdasarkan kode uniknya.",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Data voucher ditemukan."
          }
        }
      },
      "patch": {
        "summary": "Perbarui Voucher (Admin)",
        "description": "Mengubah data voucher yang sudah ada.",
        "tags": [
          "Order & Checkout"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher berhasil diperbarui."
          }
        }
      },
      "delete": {
        "summary": "Hapus Voucher (Admin)",
        "description": "Menghapus voucher dari sistem.",
        "tags": [
          "Order & Checkout"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Voucher berhasil dihapus."
          }
        }
      }
    },
    "/vouchers/check": {
      "post": {
        "summary": "Periksa Validitas Voucher",
        "description": "Mengecek apakah sebuah kode voucher bisa digunakan untuk keranjang belanja saat ini. Menghitung jumlah potongan harga.",
        "tags": [
          "Order & Checkout"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "total",
                  "items"
                ],
                "properties": {
                  "code": {
                    "type": "string",
                    "example": "SKINCARENEW"
                  },
                  "total": {
                    "type": "number",
                    "description": "Total belanja awal"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Voucher valid. Mengembalikan nilai diskon."
          },
          "400": {
            "description": "Voucher tidak valid (expired, kuota habis, atau syarat tidak terpenuhi)."
          }
        }
      }
    },
...
    "/orders": {
      "post": {
        "summary": "Buat Pesanan Baru (Checkout)",
        "description": "Mengubah isi keranjang menjadi pesanan resmi. Menghitung total biaya termasuk ongkir dan diskon.",
        "tags": [
          "Order & Checkout"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "user_id",
                  "courier",
                  "shipping_cost",
                  "shipping_service"
                ],
                "properties": {
                  "user_id": {
                    "type": "string"
                  },
                  "courier": {
                    "type": "string",
                    "example": "jne"
                  },
                  "shipping_service": {
                    "type": "string",
                    "example": "REG"
                  },
                  "shipping_cost": {
                    "type": "number",
                    "example": 15000
                  },
                  "coupon_id": {
                    "type": "string",
                    "description": "Opsional jika pakai voucher"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pesanan berhasil dibuat."
          }
        }
      },
      "get": {
        "summary": "Ambil Semua Pesanan (Admin)",
        "description": "Melihat seluruh daftar pesanan dari semua pelanggan (Khusus Admin).",
        "tags": [
          "Order & Checkout"
        ],
        "responses": {
          "200": {
            "description": "Daftar pesanan berhasil diambil."
          }
        }
      }
    },
    "/orders/{user_id}": {
      "get": {
        "summary": "Riwayat Pesanan Saya (Pelanggan)",
        "description": "Mengambil daftar pesanan milik pelanggan tertentu berdasarkan ID-nya.",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar riwayat pesanan ditemukan."
          }
        }
      }
    },
    "/orders/{id}": {
      "put": {
        "summary": "Perbarui Data Pesanan (Admin)",
        "description": "Mengubah informasi pada pesanan yang sudah ada (misal: ganti alamat, kurir, atau penyesuaian nominal).",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pesanan berhasil diperbarui."
          }
        }
      }
    },
    "/orders/{id}/status": {
      "patch": {
        "summary": "Update Status Pesanan (Admin/Sistem)",
        "description": "Mengubah status pesanan (misal: dari 'Pending' ke 'Processing').",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "processing"
                  },
                  "note": {
                    "type": "string",
                    "description": "Catatan tambahan"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status pesanan berhasil diupdate."
          }
        }
      }
    },
    "/orders/detail/{id}": {
      "get": {
        "summary": "Lihat Detail Pesanan Lengkap",
        "description": "Mengambil informasi sangat rinci tentang satu pesanan, termasuk item yang dibeli, status pembayaran, dan data pengiriman.",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail pesanan ditemukan."
          }
        }
      }
    },
    "/orders/{id}/ship": {
      "put": {
        "summary": "Input Resi Pengiriman (Admin)",
        "description": "Memasukkan nomor resi kurir dan mengubah status pesanan menjadi 'Shipped' (Dikirim).",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "tracking_number"
                ],
                "properties": {
                  "tracking_number": {
                    "type": "string",
                    "example": "JNE123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nomor resi berhasil disimpan."
          }
        }
      }
    },
    "/orders/{id}/complete": {
      "put": {
        "summary": "Konfirmasi Pesanan Selesai",
        "description": "Menandakan bahwa barang sudah diterima oleh pelanggan. Status berubah menjadi 'Completed'.",
        "tags": [
          "Order & Checkout"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pesanan dinyatakan selesai."
          }
        }
      }
    },
...
  },
...
}